+++ packages/@uppy/core/src/Uppy.test.js
